Apply Intellij quick fixes to improve code quality in debug core and ui

+++ plugins/org.eclipse.m2m.qvt.oml.debug.core/src/org/eclipse/m2m/qvt/oml/debug/core/vm/protocol/VMBreakpointResponse.java
@@ -30,7 +30,7 @@ public class VMBreakpointResponse extends VMResponse {
 		fAddedBreakpointsIDs = new long[addedBreakpointIDs.size()];
 		int i = 0;
 		for (Long id : addedBreakpointIDs) {
-			fAddedBreakpointsIDs[i++] = id.longValue();
+			fAddedBreakpointsIDs[i++] = id;
 		}
 	}
